feat(plugins): kernel: fix ABI information setting
ABI information should only be assumed when a flavour is used to generate
a config. Otherwise, version information will be encoded in
include/config/kernel.release and can be read from there. As such, either
the information is known beforehand and can be set at build-time, or it
is known after the fact and must be read post-build.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>