feat(plugins): kernel,initrd: remove dangling links
build|source are broken symlinks in a runtime context; remove them to
avoid interfering with normal functionality.

Ensure the correct ones are removed by replacing a glob with $kver, an
already well-defined value for the presumed directory name.

If they somehow manage to survive into the initrd environment, ensure
they are deleted there as well.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>