feat(plugins): initrd: remove link_files
Reimplement link_files() as a separate kind of copy for reach invoker.
The logic may not be foolproof yet (for instance, how do we handle files
which are symlinks?) but at least things are now easier to grok.

The reimplementation should be one-to-one with the old behavior in at
least the simpler and more straightforward cases.

A causualty of this operation was removing an option previously supported,
initrd-overlay. overlay is a term of art for crafts, and it's similar
enough to the addons option that I don't think splitting hairs over when
to use one over the other is meaningful enough to warrant carrying around
this behavior.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>