feat(plugins): initrd|kernel: don't error on unlink
The kernel plugin should remove the build and source symlinks in
lib/modules/kver/ as they will be broken links regardless, but just
in-case the kernel plugin behavior changes and this is no longer the case
we still want them unlinked in at least the initrd case. So don't
outright fail if the link names don't exist, just keep going.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>