feat(plugins): initrd: set a default for options
Some new options are added to the initrd plugin which aren't available
when using the legacy plugin, but it is good practice to set the variable
to a value (False, "") regardless. It's more appropriate to do in the
script because the legacy plugin doesn't implement the same schema as the
non-legacy one, and it would be cruel to provide an option which can only
be a single value.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>