feat(plugins): initrd: remove base check
Series information is acquired from the build environment rather than
passsed from the plugin itself.

The build environment must "sufficiently match" the runtime environment
anyways, so where we are building should be done correctly by snapcraft
and we should rely on it to do what the crafter is intendeding and use
the information provided to us as such.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>