feat(plugins): initrd: build initrd with features
ubuntu-core-initramfs supports building the initrd with features by
placing a directory named for the feature in
usr/lib/ubuntu-core-initramfs/{feature}. The contents of those
directories will make their way into the built initrd.

If the user specifies initrd-{addons,firmware}, the relevant feature
should be "enabled" (created, used).

Technically the relevant files could be copied into the `main/` feature,
but doing it this way it makes it more explicit what is being done with
ubuntu-core-initramfs and should ease maintenance if this behavior
changes iin the future.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>