refactor: move app models to their own file.
The project model file is untenably large, making it very hard to work
with. This splits the app model (and child models) into its own file,
helping reduce the complexity.

While it also moves the tests as appropriate, it does not (yet) modify
the app tests to only unmarshal the app model.

Signed-off-by: Alex Lowe <alex.lowe@canonical.com>