feat(plugins): initrd: improve UKI option checking
A stub only exists for some architectures and even then only starting in
certain releases.

The correct package for the stub would be installed within the chroot as
a dependency of ubuntu-core-initramfs in all environments where a UKI is
supported (provided by systemd or as a (recommended) dependency of
systemd-ukify depending on the release), so the only change we need is to
ensure the option makes sense for the target architecture.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>