feat(plugins): initrd: modify cert/key defaults
Because the snakeoil key and cert are provided by ubuntu-core-initramfs
but that is never installed on the build host and only within the initrd
chroot, it doesn't entirely "make sense" for their paths to be called
default values for the key and cert (those paths will "never exist").

As such, make the plugin consider the defaults to be empty and instead
make whether or not the snakeoil key and cert are used an implementation
detail of the scipt; a key and cert are needed whenever building a UKI
anyways, it's just that if none are provided ubuntu-core-initramfs would
always use the snakeoil anyways.

As a consequence we can remove the prep_sign function.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>