feat(plugins): initrd: modify validation logic
As providing your own key or cert only make sense if you're building a
UKI anyways, assume that the user intends to build one if they specify a
cert or key. If they don't specify a cert or key but want a UKI, the
script will correctly handle what to do.

Condense the checking logic into build command as it's a bit more compact
and straightforward what we're doing if we keep it close to the script.
If plugin options could be made to depend on the target architecture,
we'd do the Error related logic as early as possible. Since we can't do
that yet, users trying to do something unsupported will have to wait a
while to find out.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>