feat(plugins): initrd|kernel: fix spreads
A quirk in the build-for processing can cause failures if they aren't
wrapped in [].

The cross-initrd test can be simplified a bit; we don't actually need to
specify the allwinner kernel, especially for any practical purposes. As
such, it's more efficient to grab whatever generic kernel exists (and by
this point, the generic kernel should support the Nezha).

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>