ci: improve the state of some flaky tests (#5899)
Improves/fixes a number of flaky tests. I recommend reviewing per-commit.

The integration store tests are currently broken and seem to need store-side fixes, so they will be left out of this PR. The remote build tests are failing for similar reasons. The snapd failures are upstream build errors, which I have alerted the snapd team of.